backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
langchain>=0.3
langchain-openai>=0.2
langchain-pinecone>=0.2
langchain-community>=0.3
langsmith>=0.1
uvicorn[standard]>=0.30
drf-spectacular>=0.27
